BRANDING: Add Support for SSO from an auxiliary server
Accept an encrypted token from an auxiliary server to provide login
information. This permits us to store account related information, such
as passwords in a separate system. It also provides the ability to have
separate instances of App Inventor that are used by different
people. The auxiliary, or login, server will have to keep track of which
App Inventor server a particular user’s account resides on.

Note: The auxiliary server will also have to ability to remotely create
accounts on the App Inventor servers, but that functionality is not yet
provided by this commit.

Change-Id: I30600308185d7d5916ed1e1907a6d576ffb58a8e