Update OdeMessages_de.properties
I made some changes in the file. I started with the bottom line and stopped in Line 350. The last lines will be done in the next days. I found the suggestions 0, 2, 4, 5, 6, 9, 14, 21, 22, 34, 36, 37, 39, 40, 41, 42, 43, 44, 46, 59, 63, 67, 68, 71, 72, 74, 75, 76 and 77 of our table. Furthermore, the use of the German "Sie" and "Du" is inconsistent in the file. Most time, you use "Sie", but there is at least one message using "Du". Best, Matthias