Add an Id to the Gallery
When communicating with the gallery provide a Gallery ID. The Gallery ID
is string that identifies the relationship between the instance of MIT
App Inventor and the given gallery. It is used both to identify an
installation of MIT App Inventor inside of the Gallery, which when
combined with the userId within an instance, identifies a user uniquely.

It is also used to select the key for encryption and decryption of the
tokens passed between MIT App Inventor and the Gallery.

Change-Id: I85175267fc45f4f9baf8bf97885fa326eca2e652